+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * v1.2.5
+  - Bugfix: Fix infinite loop (permanent high CPU) by working around a crypto/ssh.mux.SendRequest() issue when called after closing the client.
   - Bugfix: Fix missing context cleanup for in-ssh connection attempts (resource leak)
 
 * v1.2.4